Code Refactoring and Modularization Done.
