+++ classes/CandlePatterns.py
