+++ screenipy.ini
@@ -4,7 +4,7 @@ daystolookback = 30
 duration = 1d
 minprice = 30
 maxprice = 5000
-volumeratio = 2
+volumeratio = 2.5
 consolidationpercentage = 10
 shuffle = y
 onlystagetwostocks = y
